Swift HEAD behavior change casued some CDMI operation fails
In previous Swift HEAD request, the response contains content
type setup when the object gets created. In Swift 1.0 version,
for some reason, this has been changed, so the HEAD request
does not return content-type value, it always set to text/plain
which is a problem. Swift needs to stick with the original
object content-type value. In any case, this implementation
has to be changed to make it work with Swift 1.6.