fixed the test image location problem
The test image was placed in the same directory where the test files
are, when running with py.test, the path was calculated fine. however
when use plain python to run the test, the path was not returned
correctly. The changes fixed the problem also use the os independent
separator in file path.