fix multipart mime update issue when msg is not generated by email
The multipart mime update against an object may send a request with
message generated not using the email library in this case, the
request body does not include the content type and boundary. thus
the message is considered as a bulk of stream. That is why it fails
getting the content for the object.